crane: rename audio/tts.rs to audio/tts_client.rs
Make room for the Tts trait module by renaming the existing TtsClient
stub to tts_client.rs.
